requirements.txt
streamlit
pandas
numpy
joblib
scikit-learn
tensorflow-cpu